feat: smal refactor
