+++ packages/cli/__tests__/icons/templates/svg-sprite.test.ts
@@ -1,11 +1,11 @@
-import { generateSvgSprite } from "../src/icons/generate-svg-sprite";
+import { svgSpriteTemplate } from "../../../src/icons/templates/svg-sprite";
 
 const heartIcon =
   '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path d="M12 21.35l-1.45-1.32C5.4 15.36 2 12.28 2 8.5 2 5.42 4.42 3 7.5 3c1.74 0 3.41.81 4.5 2.09C13.09 3.81 14.76 3 16.5 3 19.58 3 22 5.42 22 8.5c0 3.78-3.4 6.86-8.55 11.54L12 21.35z"/></svg>';
 
 describe("generate-svg-sprite", () => {
   it("should generate sprite", () => {
-    const result = generateSvgSprite([
+    const result = svgSpriteTemplate([
       {
         content: heartIcon,
         name: "heart",
